move some stuff out of index.ts to make it more clear that things are examples; move some util functions to utils.ts
